% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defactor.R
\name{defactor}
\alias{defactor}
\alias{defactored_osmar}
\title{Set all factor data to character}
\usage{
defactor(osmar)
}
\arguments{
\item{osmar}{An \code{\link[osmar:osmar]{osmar}} object.}
}
\value{
A \code{\link{defactored_osmar}} object with all internal tables set to use
character
}
\description{
While many internal \link[osmar:osmar]{osmar} functions rely on internal table to use
factors, it can be useful to have those data frame columns cast to character
instead. Note that several functions such as \code{\link[osmar:find]{find()}} will no longer
work on the resulting object.
}
