% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defactor.R
\name{defactor}
\alias{defactor}
\title{Set all factor data to character}
\usage{
defactor(osmar)
}
\arguments{
\item{osmar}{An \code{\link[osmar:osmar]{osmar::osmar}} object.}
}
\value{
An \code{\link[osmar:osmar]{osmar::osmar}} object with all internal tables set to use character
}
\description{
Set all factor data to character
}
