% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{read_big_osm}
\alias{read_big_osm}
\title{Read large OSM XML files}
\usage{
read_big_osm(file)
}
\arguments{
\item{file}{Path to the file}
}
\value{
An \code{\link[osmar:osmar]{osmar::osmar}} object
}
\description{
Read large OSM XML files
}
